// src/common/enum.ts
export enum UserRole {
  POSAdmin = 'POS_ADMIN',
  POSUser = 'POS_USER',
  CompanyAdmin = 'COMPANY_ADMIN',
  CompanyUser = 'COMPANY_USER',
  CompanyOwner = 'COMPANY_OWNER',
}

export enum DiscountType {
  Amount = 'AMOUNT',
  Percent = 'PERCENT',
}

export enum CompanyStatus {
  Active = 'ACTIVE',
  InActive = 'INACTIVE',
}